refactor(gateway): extract redaction/provider-error helpers from run.py (slice 9 of #54962)
